Quote, list (not listitem) behave the same
Those both act like a multiline container, closed or removed with double-enter.
