+++ aztec/src/main/kotlin/org/wordpress/aztec/handlers/GenericBlockHandler.kt
@@ -1,0 +1,38 @@
+package org.wordpress.aztec.handlers
+
+import org.wordpress.aztec.spans.AztecBlockSpan
+import org.wordpress.aztec.watchers.TextDeleter
+
+/**
+ * A general html block editing handler that closes when a newline is entered at an empty line ("double-enter").
+ * If completely empty, the whole block is removed with double-enter.
+ */
+open class GenericBlockHandler<T : AztecBlockSpan>(clazz: Class<T>) : BlockHandler<T>(clazz) {
+    // fun handleNewlineAtStartOfBlock()
+    // nothing special to do
+
+    override fun handleNewlineAtEmptyLineAtBlockEnd() {
+        // adjust the block end to only include the chars before the newline just added
+        block.end = newlineIndex
+
+        // delete the newline
+        TextDeleter.mark(text, newlineIndex, newlineIndex + 1)
+    }
+
+    override fun handleNewlineAtEmptyBody() {
+        // block is empty so, remove it
+        block.remove()
+
+        // delete the newline
+        TextDeleter.mark(text, newlineIndex, newlineIndex + 1)
+    }
+
+    // fun handleNewlineAtTextEnd()
+    // got a newline while being at the end-of-text. We'll let the block engulf it.
+
+    // fun handleNewlineInBody()
+    // newline added at some position inside the block. Nothing special to do.
+
+    // fun handleEndOfBufferMarker()
+    // nothing special to do
+}
